Marathron_website
Created a web site that lets users sign up for the fictitious SDSU Marathon. On client side JavaScript and JQuery has been used    to validate the form. While on server side, Ajax and PHP is used to read the parameters of the form stored in the DB and to generate a report that gives the roster of the runners, grouped by category (teen, adult, and senior) and alphabetized by last name.